+++ frontend/constants/api/queryParams.ts
@@ -1,4 +1,4 @@
-import { ServiceType } from '@/types/api/recruitment';
+import { ServiceType } from '@/types/api/serviceType';
 
 export const SERVICE_TYPES: Array<ServiceType> = ['산업기능요원', '전문연구요원', '승선근무예비역'] as const;
 export const SERVICE_STATUSES = ['보충역', '현역'];
