+++ frontend/types/api/experienceLevel.ts
@@ -1,0 +1,4 @@
+export interface ExperienceLevel {
+  start: '신입' | '1년' | '2년' | '3년' | '4년' | '5년';
+  end: '신입' | '1년' | '2년' | '3년' | '4년' | '5년';
+}
