+++ frontend/types/api/serviceStatus.ts
@@ -1,0 +1,1 @@
+export type ServiceStatus = '보충역' | '현역';
