+++ frontend/types/api/serviceType.ts
@@ -1,0 +1,1 @@
+export type ServiceType = '산업기능요원' | '전문연구요원' | '승선근무예비역';
